+++ nengo/networks/ensemblearray.py
@@ -223,10 +223,10 @@ class EnsembleArray(nengo.Network):
 
         .. testcode::
 
-           ea.add_output('output', None, solver=nengo.solvers.Lstsq())
+           ea.add_output("lstsq_output", None, solver=nengo.solvers.Lstsq())
 
-        creates a new output with the decoders of each connection solved for
-        with the `.Lstsq` solver.
+        creates a new output at ``ea.lstsq_output`` with the decoders
+        of each connection solved for with the `.Lstsq` solver.
 
         Parameters
         ----------
@@ -245,6 +245,14 @@ class EnsembleArray(nengo.Network):
             should stay as None, and synaptic filtering should be performed in
             the connection from the output node.
         """
+        if hasattr(self, name):
+            raise ValidationError(
+                "Cannot add output named %r; there is already an attribute by this name"
+                % (name,),
+                attr="name",
+                obj=self,
+            )
+
         dims_per_ens = self.dimensions_per_ensemble
 
         # get output size for each ensemble
@@ -253,7 +261,7 @@ class EnsembleArray(nengo.Network):
         if is_iterable(function) and all(callable(f) for f in function):
             if len(list(function)) != self.n_ensembles:
                 raise ValidationError(
-                    "Must have one function per ensemble", attr="function"
+                    "Must have one function per ensemble", attr="function", obj=self
                 )
 
             for i, func in enumerate(function):
@@ -268,6 +276,7 @@ class EnsembleArray(nengo.Network):
             raise ValidationError(
                 "'function' must be a callable, list of callables, or None",
                 attr="function",
+                obj=self,
             )
 
         output = nengo.Node(output=None, size_in=sizes.sum(), label=name)